Add config option to not record if it is an old tweet
If the tweet is more than `tweet_difference_limit` then it will
not reply to the tweet with media (assumes person has moved on)

This resolves #7

+++ TweetStreetCam.py
@@ -57,52 +57,73 @@ class GraffCam:
 		script_graffcam = self.script_graffcam
 		script_ta = self.script_ta
 
-
+		# Work out time difference of tweet & get tweet ID
 		last_mention_id = tweet['id']
+		tweet_time = datetime.datetime.strptime(tweet['created_at'],'%a %b %d %H:%M:%S +0000 %Y')
+		current_time = datetime.datetime.strptime(time.strftime('%Y-%m-%d %H:%M:%S'), "%Y-%m-%d %H:%M:%S")
+		time_tweet_difference = (current_time - tweet_time).seconds / 60
 
 		# Make a user
 		user = tweet['user']
 		username = user['screen_name']
 
+		# If the username is not itself
 		if username != self.config.get('setup', 'twitter_username'):
-			if self._DEBUG_MODE == 'False':
+
+			# If it's bigger than config option - don't record
+			if int(time_tweet_difference) <= int(self.config.get('options', 'tweet_difference_limit')):
+
+				# Get a prerpation tweet text
 				start_status = random.choice(json.loads(self.config.get('tweet_text', 'preperation')))
 				start_status = start_status.replace('[[user]]', '@%s' % (username))
-				start_post = self.api.request('statuses/update', {'status': start_status, 'in_reply_to_status_id': tweet['id']})
-				time.sleep(5)
+
+				if self._DEBUG_MODE == 'False':
+					start_post = self.api.request('statuses/update', {'status': start_status, 'in_reply_to_status_id': tweet['id']})
+					time.sleep(5)
+				else:
+					print 'Preperation tweet: %s' % (start_status)
+
+				#Initialise custom classes
+				graffcam = Graffcam(self._HOME_PATH, self.camera, script_graffcam)
+				ta = TA(self._HOME_PATH, self.api, script_ta)
+
+				# If the tweet contains a photo trigger hashtag
+				if ta.is_photo(tweet):
+					media = graffcam.capture_photo(tweet)
+					media_upload = ta.upload_image(media)
+					status_pick = self.config.get('tweet_text', 'photo')
+				else:
+					media = graffcam.record_video(tweet)
+					media_upload = ta.upload_video(media)
+					status_pick = self.config.get('tweet_text', 'video')
+
+				# Build the status and send
+				status = random.choice(json.loads(status_pick))
+				status = status.replace('[[user]]', '@%s' % (username))
+
+				if self._DEBUG_MODE == 'False':
+					if media_upload.status_code > 199 or media_upload.status_code < 300:
+						post = self.api.request('statuses/update', {'status': status, 'in_reply_to_status_id': tweet['id'], 'media_ids': media_upload.json()['media_id']})
+						if post.status_code > 199 or post.status_code < 300:
+							os.remove(media)
+				else:
+					print 'Original tweet: %s' % (tweet['text'])
+					print 'Status: %s [media: %s] ' % (status, media)
 			else:
-				print 'Preperation tweet: %s' % (start_status)
 
-			#Initialise custom classes
-			graffcam = Graffcam(self._HOME_PATH, self.camera, script_graffcam)
-			ta = TA(self._HOME_PATH, self.api, script_ta)
+				# Prepare the late text
+				late_status = random.choice(json.loads(self.config.get('tweet_text', 'late')))
+				late_status = late_status.replace('[[user]]', '@%s' % (username))
 
-			# If the tweet contains a photo trigger hashtag
-			if ta.is_photo(tweet):
-				media = graffcam.capture_photo(tweet)
-				media_upload = ta.upload_image(media)
-				status_pick = self.config.get('tweet_text', 'photo')
-			else:
-				media = graffcam.record_video(tweet)
-				media_upload = ta.upload_video(media)
-				status_pick = self.config.get('tweet_text', 'video')
+				if self._DEBUG_MODE == 'False':
+					late_post = self.api.request('statuses/update', {'status': late_status, 'in_reply_to_status_id': tweet['id']})
+				else:
+					print 'Tweet was late (%s minutes): %s' % (time_tweet_difference, late_status)
 
-			# Build the status and send
-			status = random.choice(json.loads(status_pick))
-			status = status.replace('[[user]]', '@%s' % (username))
-
+			# Update the last ID
 			if self._DEBUG_MODE == 'False':
-				if media_upload.status_code > 199 or media_upload.status_code < 300:
-					post = self.api.request('statuses/update', {'status': status, 'in_reply_to_status_id': tweet['id'], 'media_ids': media_upload.json()['media_id']})
-					if post.status_code > 199 or post.status_code < 300:
-						os.remove(media)
-			else:
-				print 'Original tweet: %s' % (tweet['text'])
-				print 'Status: %s [media: %s] ' % (status, media)
+				self.config.set('tweets', 'last_mention_id', last_mention_id)
+				with open(self._HOME_PATH + '_config.cfg', 'w') as f:
+					self.config.write(f)
 		else:
 			print 'Tweet by Graffcam, not actioning anything'
-		# Update the last ID
-		if self._DEBUG_MODE == 'False':
-			self.config.set('tweets', 'last_mention_id', last_mention_id)
-			with open(self._HOME_PATH + '_config.cfg', 'w') as f:
-				self.config.write(f)
